# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Application API",
        "description": "API Documentation for Online Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/addToCart/{product_id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "إضافة منتج للكارت",
                "operationId": "39b01a344f6996cb26ef7c291b61ba55",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 101
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم الإضافة للكارت بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم الإضافة للكارت بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "المنتج out of stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "المنتج ده out of stock"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getCart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "جلب محتويات الكارت",
                "operationId": "e685f97e7c56857a0acdfb49ac1c1d32",
                "responses": {
                    "200": {
                        "description": "بيانات الكارت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "image": {
                                                                "type": "string",
                                                                "example": "http://example.com/storage/products/img.jpg",
                                                                "nullable": true
                                                            },
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 101
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "منتج A"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "unit_price": {
                                                                "type": "number",
                                                                "example": 50
                                                            },
                                                            "unit_tax": {
                                                                "type": "number",
                                                                "example": 7.5
                                                            },
                                                            "unit_price_after_tax": {
                                                                "type": "number",
                                                                "example": 57.5
                                                            },
                                                            "total_price": {
                                                                "type": "number",
                                                                "example": 115
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "number_of_products": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "final_price": {
                                                    "type": "number",
                                                    "example": 345
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/removeFromCart/{product_id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "حذف منتج من الكارت",
                "operationId": "4114daae56e47dfb0d4b5b1bba78216e",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 101
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم مسح المنتج من الكارت بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم مسح المنتج من الكارت بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "المنتج مش موجود في الكارت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "المنتج مش موجود في الكارت"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/addToFavourites/{product_id}": {
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "إضافة منتج للمفضلة",
                "operationId": "964258ee78b40a1cd32e074cd8ae3deb",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 101
                    }
                ],
                "responses": {
                    "201": {
                        "description": "تم الإضافة للمفضلة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم الإضافة للمفضلة بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "المنتج موجود بالفعل في المفضلة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "المنتج موجود بالفعل في المفضلة"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getFavourites": {
            "get": {
                "tags": [
                    "Favourites"
                ],
                "summary": "جلب المفضلة",
                "operationId": "feee21e64ccd9257c7a291e9e7154ea4",
                "responses": {
                    "200": {
                        "description": "قائمة المفضلة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://example.com/storage/products/img.jpg",
                                                        "nullable": true
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "منتج A"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 57.5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}